iot-api/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.22
pymysql==1.1.0
paho-mqtt==1.6.1
